# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InfraPulse MSK",
    "description": "API InfraPulse MSK. Документация открыта для чтения; каждый маршрут, кроме `/health/*` и входа, проверяет сессию или токен интеграции.\n\n* **Чтение (GET)** — по cookie сессии: войти в веб-интерфейс этого же сайта, затем «Try it out» здесь.\n* **Пачки наблюдений** `POST /api/v1/observations` — токен интеграции: «Authorize» → `integrationToken`, значение `ipk_...` без слова Bearer.\n* **Запись под сессией (POST)** — дополнительно `csrfToken`: значение cookie `__Host-infrapulse-csrf`.\n\nИнструкция для систем-источников — `docs/INTEGRATION_API.md` в репозитории.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/imports": {
      "post": {
        "summary": "Upload",
        "operationId": "upload_api_v1_imports_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_imports_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportFile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      },
      "get": {
        "summary": "Imports",
        "operationId": "imports_api_v1_imports_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/v1/imports/{import_id}": {
      "get": {
        "summary": "Import File",
        "operationId": "import_file_api_v1_imports__import_id__get",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportFile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_api_v1_imports_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "format": {
            "type": "string",
            "enum": [
              "journal_csv",
              "journal_json",
              "reference_channels_csv",
              "reference_objects_csv",
              "reference_states_csv"
            ],
            "title": "Format",
            "default": "journal_csv"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_imports_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportFile": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Id"
          },
          "format": {
            "type": "string",
            "enum": [
              "journal_csv",
              "journal_json",
              "reference_channels_csv",
              "reference_objects_csv",
              "reference_states_csv"
            ],
            "title": "Format"
          },
          "file_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "File Name"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          },
          "size_bytes": {
            "type": "integer",
            "maximum": 52428800.0,
            "minimum": 0.0,
            "title": "Size Bytes"
          },
          "uploaded_by": {
            "type": "string",
            "minLength": 1,
            "title": "Uploaded By"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "parsing",
              "imported",
              "recomputing",
              "published",
              "duplicate",
              "failed"
            ],
            "title": "Status"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "rows_total": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Total"
          },
          "rows_accepted": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Accepted"
          },
          "rows_duplicate": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Duplicate"
          },
          "rows_quarantined": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Quarantined"
          },
          "unknown_channels": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Unknown Channels"
          },
          "event_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event From"
          },
          "event_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event To"
          },
          "reference_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Version"
          },
          "duplicate_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of"
          },
          "forecast_generation": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Generation"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "file_too_large",
                  "unknown_format",
                  "bad_header",
                  "bad_encoding",
                  "no_valid_rows",
                  "reference_missing",
                  "recompute_failed",
                  "internal_error"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "timings": {
            "items": {
              "$ref": "#/components/schemas/ImportStageTiming"
            },
            "type": "array",
            "title": "Timings"
          },
          "quarantine_reasons": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "bad_column_count",
                "bad_date",
                "bad_time",
                "bad_bool",
                "empty_channel",
                "bad_encoding",
                "value_too_long",
                "channel_type_conflict"
              ]
            },
            "type": "object",
            "title": "Quarantine Reasons"
          },
          "quarantine_sample": {
            "items": {
              "$ref": "#/components/schemas/QuarantineSample"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Quarantine Sample"
          },
          "new_card_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "New Card Ids"
          },
          "released_card_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Released Card Ids"
          },
          "source_layout": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "organizers_export",
                  "tz_appendix1",
                  "api_batch",
                  "reference"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Layout"
          },
          "source_container": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "csv",
                  "xlsx",
                  "json",
                  "xml"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Container"
          },
          "alarm_not_provided": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Alarm Not Provided"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Notes"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "format",
          "file_name",
          "sha256",
          "size_bytes",
          "uploaded_by",
          "uploaded_at",
          "status"
        ],
        "title": "ImportFile",
        "description": "One uploaded file and its processing report.\n\nRow counts: ``rows_total = rows_accepted + rows_duplicate + rows_quarantined``.\n``unknown_channels`` are channels absent from the reference; their rows are kept\nwith an unknown object (DATA-04). ``forecast_generation`` is the publication that\nfirst included the file."
      },
      "ImportList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ImportFile"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit"
        ],
        "title": "ImportList"
      },
      "ImportStageTiming": {
        "properties": {
          "stage": {
            "type": "string",
            "enum": [
              "store",
              "parse",
              "load",
              "detect",
              "score",
              "publish"
            ],
            "title": "Stage"
          },
          "seconds": {
            "type": "number",
            "minimum": 0.0,
            "title": "Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "stage",
          "seconds"
        ],
        "title": "ImportStageTiming"
      },
...
      "QuarantineSample": {
        "properties": {
          "line_no": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Line No"
          },
          "reason": {
            "type": "string",
            "enum": [
              "bad_column_count",
              "bad_date",
              "bad_time",
              "bad_bool",
              "empty_channel",
              "bad_encoding",
              "value_too_long",
              "channel_type_conflict"
            ],
            "title": "Reason"
          },
          "raw_excerpt": {
            "type": "string",
            "maxLength": 200,
            "title": "Raw Excerpt"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "line_no",
          "reason",
          "raw_excerpt"
        ],
        "title": "QuarantineSample",
        "description": "A rejected line with its reason; the raw excerpt is shown only to administrators."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "integrationToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ipk_...",
        "description": "Токен интеграции внешней системы (выдаёт администратор сервиса). Даёт только право ingest: отправку пачек и статус своих пачек."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-infrapulse-session",
        "description": "Сессия входа через каталог. Ставится при входе в веб-интерфейс этого же сайта и отправляется браузером сама; поле в «Authorize» заполнять не нужно."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Для POST под сессией: значение cookie __Host-infrapulse-csrf. С токеном интеграции не нужен."
      }
    }
  }
}